+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -4,6 +4,6 @@ package ru.skypro.homework.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.homework.model.AdEntity;
 
-public interface AdRepository extends JpaRepository<AdEntity, Integer> {
+public interface AdRepository extends JpaRepository<AdEntity, Long> {
 
 }
