+++ src/main/java/ru/skypro/homework/repository/CommentRepository.java
@@ -5,7 +5,7 @@ import ru.skypro.homework.model.CommentEntity;
 
 import java.util.List;
 
-public interface CommentRepository extends JpaRepository<CommentEntity, Integer> {
+public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
 //    List<CommentEntity> findByAdId(Integer adId);
 //
 //    void deleteByAdIdAndId(Integer adId, Integer commentId);
